+++ node.cpp
@@ -7,11 +7,11 @@ node::node(int id) {
 }
 
 
-void node::addUnary(float value) {
+void node::addUnary(double value) {
    unaries[next_unary] = value;
    ++next_unary;
 }
 
-float node::get_min_unary() {
+double node::get_min_unary() {
     return *min_element(unaries.begin(), unaries.end());
 }
